requirements.txt
playwright==1.41.0
pytest==7.4.3
pytest-playwright==0.4.4
pytest-html==4.1.1
pytest-xdist==3.5.0
requests==2.31.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
python-dotenv==1.0.0
jinja2==3.1.3
colorama==0.4.6
